+++ src/main/java/roomescape/business/service/ReservationService.java
@@ -12,12 +12,12 @@ import roomescape.business.model.entity.Theme;
 import roomescape.business.model.repository.ReservationDao;
 import roomescape.business.model.repository.ReservationTimeDao;
 import roomescape.business.model.repository.ThemeDao;
+import roomescape.exception.impl.EntityNotFoundException;
+import roomescape.exception.impl.PastReservationException;
+import roomescape.exception.impl.ReservationTimeConflictException;
 import roomescape.presentation.dto.request.ReservationRequest;
 import roomescape.presentation.dto.response.AvailableReservationTimeResponse;
 import roomescape.presentation.dto.response.ReservationResponse;
-import roomescape.exception.EntityNotFoundException;
-import roomescape.exception.PastReservationException;
-import roomescape.exception.ReservationTimeConflictException;
 
 @Service
 public class ReservationService {
