+++ src/main/java/roomescape/business/service/ReservationTimeService.java
@@ -5,11 +5,11 @@ import org.springframework.stereotype.Service;
 import roomescape.business.model.entity.ReservationTime;
 import roomescape.business.model.repository.ReservationDao;
 import roomescape.business.model.repository.ReservationTimeDao;
+import roomescape.exception.impl.EntityNotFoundException;
+import roomescape.exception.impl.ReservationExistException;
+import roomescape.exception.impl.ReservationTimeExistException;
 import roomescape.presentation.dto.request.ReservationTimeRequest;
 import roomescape.presentation.dto.response.ReservationTimeResponse;
-import roomescape.exception.EntityNotFoundException;
-import roomescape.exception.ReservationExistException;
-import roomescape.exception.ReservationTimeExistException;
 
 @Service
 public class ReservationTimeService {
