--- src/main/java/roomescape/exception/DuplicateReservationException.java
@@ -1,7 +1,0 @@
-package roomescape.exception;
-
-public class DuplicateReservationException extends RoomescapeException {
-    public DuplicateReservationException(String message) {
-        super(message);
-    }
-}
