+++ src/main/java/roomescape/exception/impl/EntityNotFoundException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.exception.impl;
+
+import roomescape.exception.RoomescapeException;
 
 public class EntityNotFoundException extends RoomescapeException {
     public EntityNotFoundException(String message) {
