+++ src/main/java/roomescape/exception/impl/PastReservationException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.exception.impl;
+
+import roomescape.exception.RoomescapeException;
 
 public class PastReservationException extends RoomescapeException {
     public PastReservationException(String message) {
