+++ src/main/java/roomescape/exception/impl/ReservationExistException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.exception.impl;
+
+import roomescape.exception.RoomescapeException;
 
 public class ReservationExistException extends RoomescapeException {
     public ReservationExistException(String message) {
