+++ src/main/java/roomescape/exception/impl/ReservationTimeConflictException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.exception.impl;
+
+import roomescape.exception.RoomescapeException;
 
 public class ReservationTimeConflictException extends RoomescapeException {
     public ReservationTimeConflictException(String message) {
