+++ src/main/java/roomescape/exception/impl/ReservationTimeExistException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.exception.impl;
+
+import roomescape.exception.RoomescapeException;
 
 public class ReservationTimeExistException extends RoomescapeException {
     public ReservationTimeExistException(String message) {
