+++ src/main/java/roomescape/exception/impl/ThemeExistException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.exception.impl;
+
+import roomescape.exception.RoomescapeException;
 
 public class ThemeExistException extends RoomescapeException {
     public ThemeExistException(String message) {
