+++ src/main/java/roomescape/presentation/api/ReservationController.java
@@ -5,7 +5,6 @@ import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,9 +16,6 @@ import roomescape.business.service.ReservationService;
 import roomescape.presentation.dto.request.ReservationRequest;
 import roomescape.presentation.dto.response.AvailableReservationTimeResponse;
 import roomescape.presentation.dto.response.ReservationResponse;
-import roomescape.exception.DuplicateReservationException;
-import roomescape.exception.PastReservationException;
-import roomescape.exception.ReservationTimeConflictException;
 
 @RestController
 @RequestMapping("/reservations")
