+++ src/main/java/roomescape/presentation/api/ThemeController.java
@@ -5,7 +5,6 @@ import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,8 +14,6 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.business.service.ThemeService;
 import roomescape.presentation.dto.request.ThemeRequest;
 import roomescape.presentation.dto.response.ThemeResponse;
-import roomescape.exception.ReservationExistException;
-import roomescape.exception.ThemeExistException;
 
 @RestController
 @RequestMapping("/themes")
